# api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: 0.0.1
  title: Monitoring API
host: 'localhost:10005'
basePath: /mon
schemes:
  - http
consumes:
  - application/json
  - multipart/form-data
produces:
  - application/json
  - text/plain
paths:
...
  /logs/count:
    x-swagger-router-controller: controller 
    get:
      description: returns the count of logs
      operationId: logCount
      parameters:
        - name: authorization
          in: header
          type: string
          description: The JWT token for req. validation
        - name: filter
          in: query
          type: string
          description: Filter records based on certain fields  
      responses:
        '200':
          description: count of services
        '400':
          description: Bad parameters
        '500':
          description: Internal server error
  /appCenter/{id}/logs:
    x-swagger-router-controller: controller
    get:
      description: Retrieve a list of logs of  service
      operationId: appCenterLog
      parameters:
        - name: authorization
          in: header
          type: string
          description: The JWT token for req. validation
        - name: page
          in: query
          type: integer
          description: Page number of the request
        - name: count
          in: query
          type: integer
          description: Number of records per page
        - name: filter
          in: query
          type: string
          description: Filter records based on certain fields
        - name: select
          in: query
          type: string
          description: Comma seperated fields to be displayed
        - name: sort
          in: query
          type: string
          description: sort parameter
        - name: id
          in: path
          type: string
          required: true
          description: Id of the object to be viewed
      responses:
        '200':
          description: List of the entites
        '400':
          description: Bad parameters
        '404':
          description: No records to list with the given parameter set.
        '500':
          description: Internal server error
  /appCenter/{id}/logs/count:
    x-swagger-router-controller: controller 
    get:
      description: returns the count of logs 
      operationId: appCenterLogCount
      parameters:
        - name: authorization
          in: header
          type: string
          description: The JWT token for req. validation
        - name: filter
          in: query
          type: string
          description: Filter records based on certain fields 
        - name: id
          in: path
          type: string
          required: true
          description: Id of the object to be viewed
      responses:
        '200':
          description: count of services
        '400':
          description: Bad parameters
        '500':
          description: Internal server error
...
  /appCenter/{srvcid}/audit/purge:
    x-swagger-router-controller: controller 
    delete:
      description: audit Purge
      operationId: appcenterAuditPurge
      parameters:
        - name: authorization
          in: header
          type: string
          description: The JWT token for req. validation
        - name: srvcid
          in: path
          type: string
          description: Service Id
          required: true  
      responses:
        '200':
          description: count of services
        '400':
          description: Bad parameters
        '500':
          description: Internal server error 
  /appCenter/{srvcid}/log/purge:
    x-swagger-router-controller: controller 
    delete:
      description: logs Purge
      operationId: logPurge
      parameters:
        - name: authorization
          in: header
          type: string
          description: The JWT token for req. validation
        - name: srvcid
          in: path
          type: string
          description: Service Id
          required: true  
      responses:
        '200':
          description: count of services
        '400':
          description: Bad parameters
        '500':
          description: Internal server error 
  /appCenter/{srvcid}/preHook:
    x-swagger-router-controller: controller
    get:
      description: Retrieve a list of audits
      operationId: preHook
      parameters:
        - name: authorization
          in: header
          type: string
          description: The JWT token for req. validation
        - name: srvcid
          in: path
          type: string
          description: Service Id
          required: true
        - name: page
          in: query
          type: integer
          description: Page number of the request
        - name: count
          in: query
          type: integer
          description: Number of records per page
        - name: filter
          in: query
          type: string
          description: Filter records based on certain fields
        - name: select
          in: query
          type: string
          description: Comma seperated fields to be displayed
        - name: sort
          in: query
          type: string
          description: sort parameter
      responses:
        '200':
          description: List of the entites
        '400':
          description: Bad parameters
        '404':
          description: No records to list with the given parameter set.
        '500':
          description: Internal server error
...
  /appCenter/{srvcid}/postHook:
    x-swagger-router-controller: controller
    get:
      description: Retrieve a list of audits
      operationId: postHook
      parameters:
        - name: authorization
          in: header
          type: string
          description: The JWT token for req. validation
        - name: srvcid
          in: path
          type: string
          description: Service Id
          required: true
        - name: page
          in: query
          type: integer
          description: Page number of the request
        - name: count
          in: query
          type: integer
          description: Number of records per page
        - name: filter
          in: query
          type: string
          description: Filter records based on certain fields
        - name: select
          in: query
          type: string
          description: Comma seperated fields to be displayed
        - name: sort
          in: query
          type: string
          description: sort parameter
      responses:
        '200':
          description: List of the entites
        '400':
          description: Bad parameters
        '404':
          description: No records to list with the given parameter set.
        '500':
          description: Internal server error
...
  /author/userRoleAudit:
    x-swagger-router-controller: controller
    get:
      description: Retrieve a list of audits
      operationId: userRoleAudit
      parameters:
        - name: authorization
          in: header
          type: string
          description: The JWT token for req. validation
        - name: page
          in: query
          type: integer
          description: Page number of the request
        - name: count
          in: query
          type: integer
          description: Number of records per page
        - name: filter
          in: query
          type: string
          description: Filter records based on certain fields
        - name: select
          in: query
          type: string
          description: Comma seperated fields to be displayed
        - name: sort
          in: query
          type: string
          description: sort parameter
      responses:
        '200':
          description: List of the entites
        '400':
          description: Bad parameters
        '404':
          description: No records to list with the given parameter set.
        '500':
          description: Internal server error
  /author/userRoleAudit/count:
    x-swagger-router-controller: controller 
    get:
      description: returns the count of audit
      operationId: userRoleAuditCount
      parameters:
        - name: authorization
          in: header
          type: string
          description: The JWT token for req. validation
        - name: filter
          in: query
          type: string
          description: Filter records based on certain fields 
      responses:
        '200':
          description: count of services
        '400':
          description: Bad parameters
        '500':
          description: Internal server error  
  /author/userDomainAudit:
    x-swagger-router-controller: controller
    get:
      description: Retrieve a list of audits
      operationId: userDomainAudit
      parameters:
        - name: authorization
          in: header
          type: string
          description: The JWT token for req. validation
        - name: page
          in: query
          type: integer
          description: Page number of the request
        - name: count
          in: query
          type: integer
          description: Number of records per page
        - name: filter
          in: query
          type: string
          description: Filter records based on certain fields
        - name: select
          in: query
          type: string
          description: Comma seperated fields to be displayed
        - name: sort
          in: query
          type: string
          description: sort parameter
      responses:
        '200':
          description: List of the entites
        '400':
          description: Bad parameters
        '404':
          description: No records to list with the given parameter set.
        '500':
          description: Internal server error
  /author/userDomainAudit/count:
    x-swagger-router-controller: controller 
    get:
      description: returns the count of audit
      operationId: userDomainAuditCount
      parameters:
        - name: authorization
          in: header
          type: string
          description: The JWT token for req. validation
        - name: filter
          in: query
          type: string
          description: Filter records based on certain fields 
      responses:
        '200':
          description: count of services
        '400':
          description: Bad parameters
        '500':
          description: Internal server error
  /author/userAudit:
    x-swagger-router-controller: controller
    get:
      description: Retrieve a list of audits
      operationId: userAudit
      parameters:
        - name: authorization
          in: header
          type: string
          description: The JWT token for req. validation
        - name: page
          in: query
          type: integer
          description: Page number of the request
        - name: count
          in: query
          type: integer
          description: Number of records per page
        - name: filter
          in: query
          type: string
          description: Filter records based on certain fields
        - name: select
          in: query
          type: string
          description: Comma seperated fields to be displayed
        - name: sort
          in: query
          type: string
          description: sort parameter
      responses:
        '200':
          description: List of the entites
        '400':
          description: Bad parameters
        '404':
          description: No records to list with the given parameter set.
        '500':
          description: Internal server error
  /author/userAudit/count:
    x-swagger-router-controller: controller 
    get:
      description: returns the count of audit
      operationId: userAuditCount
      parameters:
        - name: authorization
          in: header
          type: string
          description: The JWT token for req. validation
        - name: filter
          in: query
          type: string
          description: Filter records based on certain fields 
      responses:
        '200':
          description: count of services
        '400':
          description: Bad parameters
        '500':
          description: Internal server error 
...
  /author/audit/{txnId}:
    x-swagger-router-controller: controller
    get:
      description: Retrieve a list of audits
      operationId: txnAudit
      parameters:
        - name: select
          in: query
          type: string
          description: Comma seperated fields to be displayed
        - name: txnId
          in: path
          type: string
          required: true
          description: Id of the object to be updated
        - name: authorization
          in: header
          type: string
          description: The JWT token for req.validation
      responses:
        '200':
          description: List of the entites
        '400':
          description: Bad parameters
        '404':
          description: No records to list with the given parameter set.
        '500':
          description: Internal server error
...